+++ CHANGES.md
@@ -3,6 +3,8 @@ v0.1.2
 + fix bug where slivar expr would segfault if labels were re-used (now quits with error message)
 + better help in `slivar make-gnotate`
 + improve and document output format for `duo-del` https://github.com/brentp/slivar/wiki/finding-deletions-in-parent-child-duos
++ add a prelude function with hasSample(INFO, key, sample) to see if sample has been added to info field by previous filter.
++ allow outputting summary table to file with `SLIVAR_SUMMARY_FILE` environment variable
 
 v0.1.1
 ======